install/stage0: change file mode bits to several fils/dirs.
Changes made:
Add `set-group-ID` bit to the directorys installed by `rkt install`
so that later the files created in those directory can be read by
users in the `rkt` group if the group exits. All the directories are
read-only by `rkt` group except for the `cas/db` which enables `rkt`
group to do database query transaction.

Change the permission bits of ppid, pod manifest, image manifest,
dropped the execution bit.

Change the gid of `stage1/rootfs` to the gid of `rkt` group if the
group exits.

Change the gid of `rootfs/rkt` directory and all its sub-directories
(`rootfs/rkt/status`, `rootfs/rkt/env` currently) to the gid of `rkt`
group if the group exits. Also add `set-group-ID` bit to them so that
files created by the pod in these directories later can be read by
process in `rkt` group. (e.g. exit code of the apps)