common, stage1: make rkt work with a systemd v227 stage1
After v227, systemd-nspawn mounts /sys inside the container as a tmpfs.
Then it mounts the real sysfs on /sys/full and bind-mounts a white list
of subsystems to /sys. This allows to controll the /sys subsystems in
the container.

We were mounting /sys inside the container RO and as a sysfs, so
systemd-nspawn couldn't mkdir its /sys/full directory and it failed.

To fix this, we now mount /sys inside the container as a tmpfs and
remove the RO remount. systemd-nspawn will do its /sys/full magic and
remount /sys as RO making everything work again.