rkt: show prepared/created/exited times in rkt list
To get the pod preparation/creation time we use the modification time in
`/var/lib/rkt/pods/*/<UUID>`. It gets modified when the pod is prepared
and when the pod is created and doesn't change afterwards.

To get the pod exit time we use the modification time in
`/var/lib/rkt/pods/*/<UUID>/stage1`. The reason is that systemd-nspawn
creates a lock file there, and removes it when the container exits (even
if it was killed).